requirements.txt
python-telegram-bot==21.6
yt-dlp>=2025.1.1
ffmpeg-python>=0.2.0
python-dotenv>=1.0.1
